+++ Packages/EurofurenceKit/Sources/EurofurenceKit/EurofurenceModel.swift
@@ -80,94 +80,6 @@ public class EurofurenceModel: ObservableObject {
     
 }
 
-// MARK: - Model Configuration
-
-extension EurofurenceModel {
-    
-    /// Represents a collection of configurable attributes the model should use during runtime.
-    public struct Configuration {
-        
-        /// Designates the intended usage environment of the model.
-        public enum Environment {
-            
-            /// The contents of the model should persist between lifecycles.
-            case persistent
-            
-            /// The contents of the model should be discarded once the model has been deallocated.
-            case memory
-            
-            fileprivate func configure(
-                persistentContainer: EurofurencePersistentContainer,
-                properties: EurofurenceModelProperties
-            ) {
-                switch self {
-                case .persistent:
-                    persistentContainer.attachPersistentStore(properties: properties)
-                    
-                case .memory:
-                    persistentContainer.attachMemoryStore()
-                }
-            }
-            
-        }
-        
-        let persistentContainer: EurofurencePersistentContainer
-        let properties: EurofurenceModelProperties
-        let keychain: Keychain
-        let api: EurofurenceAPI
-        let conventionIdentifier: ConventionIdentifier
-        
-        private static func versionedAPI(for conventionIdentifier: ConventionIdentifier) -> EurofurenceAPI {
-            let configuration = CIDSensitiveEurofurenceAPI.Configuration(
-                conventionIdentifier: conventionIdentifier.stringValue,
-                hostVersion: "4.0.0"
-            )
-            
-            return CIDSensitiveEurofurenceAPI(configuration: configuration)
-        }
-        
-        public init(
-            environment: Environment = .persistent,
-            properties: EurofurenceModelProperties = AppGroupModelProperties.shared,
-            keychain: Keychain = SecKeychain.shared,
-            conventionIdentifier: ConventionIdentifier = .current
-        ) {
-            let apiConfiguration = CIDSensitiveEurofurenceAPI.Configuration(
-                conventionIdentifier: conventionIdentifier.stringValue,
-                hostVersion: "4.0.0"
-            )
-            
-            let api = CIDSensitiveEurofurenceAPI(configuration: apiConfiguration)
-            
-            self.init(
-                environment: environment,
-                properties: properties,
-                keychain: keychain,
-                api: api,
-                conventionIdentifier: conventionIdentifier
-            )
-        }
-        
-        public init(
-            environment: Environment = .persistent,
-            properties: EurofurenceModelProperties = AppGroupModelProperties.shared,
-            keychain: Keychain = SecKeychain.shared,
-            api: EurofurenceAPI,
-            conventionIdentifier: ConventionIdentifier = .current
-        ) {
-            self.persistentContainer = EurofurencePersistentContainer()
-            self.properties = properties
-            self.keychain = keychain
-            self.api = api
-            self.conventionIdentifier = conventionIdentifier
-            
-            environment.configure(persistentContainer: persistentContainer, properties: properties)
-        }
-        
-    }
-    
-}
-
 // MARK: - Cloud Status
 
 extension EurofurenceModel {
@@ -365,72 +277,10 @@ extension EurofurenceModel {
         do {
             try await signOut()
         } catch {
-            logger.error("Failed to automatically sign out user.", metadata: ["Error": .string(String(describing: error))])
-        }
-    }
-    
-}
-        
-// MARK: - Fetching Entities
-
-extension EurofurenceModel {
-    
-    /// Fetches the `Announcement` associated with the given identifier.
-    ///
-    /// - Parameter identifier: The identifier of the announcement to be fetched.
-    /// - Returns: The `Announcement` associated with the given identifier.
-    /// - Throws: `EurofurenceError.invalidAnnouncement` if no `Announcement` is associated with the given identifier.
-    public func announcement(identifiedBy identifier: String) throws -> Announcement {
-        try entity(identifiedBy: identifier, throwWhenMissing: .invalidAnnouncement(identifier))
-    }
-    
-    /// Fetches the `Event` associated with the given identifier.
-    ///
-    /// - Parameter identifier: The identifier of the event to be fetched.
-    /// - Returns: The `Event` associated with the given identifier.
-    /// - Throws: `EurofurenceError.invalidEvent` if no `Event` is associated with the given identifier.
-    public func event(identifiedBy identifier: String) throws -> Event {
-        try entity(identifiedBy: identifier, throwWhenMissing: .invalidEvent(identifier))
-    }
-    
-    /// Fetches the `Dealer` associated with the given identifier.
-    ///
-    /// - Parameter identifier: The identifier of the dealer to be fetched.
-    /// - Returns: The `Dealer` associated with the given identifier.
-    /// - Throws: `EurofurenceError.invalidDealer` if no `Dealer` is associated with the given identifier.
-    public func dealer(identifiedBy identifier: String) throws -> Dealer {
-        try entity(identifiedBy: identifier, throwWhenMissing: .invalidDealer(identifier))
-    }
-    
-    /// Fetches the `Message` associated with the given identifier.
-    ///
-    /// - Parameter identifier: The identifier of the message to be fetched.
-    /// - Returns: The `Message` associated with the given identifier.
-    /// - Throws: `EurofurenceError.invalidMessage` if no `Message` is associated with the given identifier.
-    public func message(identifiedBy identifier: String) throws -> Message {
-        let fetchRequest: NSFetchRequest<Message> = Message.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "identifier == %@", identifier)
-        fetchRequest.fetchLimit = 1
-        
-        let fetchResults = try viewContext.fetch(fetchRequest)
-        if let message = fetchResults.first {
-            return message
-        } else {   
-            throw EurofurenceError.invalidMessage(identifier)
-        }
-    }
-    
-    private func entity<E>(
-        identifiedBy identifier: String,
-        throwWhenMissing: @autoclosure () -> EurofurenceError
-    ) throws -> E where E: Entity {
-        let fetchRequest: NSFetchRequest<E> = E.fetchRequestForExistingEntity(identifier: identifier)
-        
-        let results = try viewContext.fetch(fetchRequest)
-        if let entity = results.first {
-            return entity
-        } else {
-            throw throwWhenMissing()
+            logger.error(
+                "Failed to automatically sign out user.",
+                metadata: ["Error": .string(String(describing: error))]
+            )
         }
     }
