+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Properties/AppGroupModelProperties.swift
