+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Properties/EurofurenceModelProperties.swift
